test(eval): move train-path wandb guard out of the entrypoint module
The fast `test_train_surge_experiment_composes_null_ckpt_without_wandb_resolution`
guard composed `train.yaml` via `initialize_config_module`, which the entrypoint
invariant `tests/_meta/test_entrypoint_e2e_only.py` forbids in `test_train.py`
(config-composition tests belong outside the canonical entrypoint modules). That
reddened the ubuntu/conda/macos jobs.

Relocate the guard to `tests/test_wandb_resolver.py` (a config-layer test module,
not an entrypoint) where the same `compose`/`_fake_api` machinery already lives,
and parametrize it over all 8 wired surge experiments. `test_train.py` keeps only
the live resume round-trip (fixture-driven, no config-initializer import).

Refs #128
